requirements.txt
beautifulsoup4
lxml
requests
datetime
psycopg2-binary
